修正key通过单元测试

+++ src/Sino.Extensions.Redis/Utils/Exit.cs
@@ -10,10 +10,10 @@ namespace Sino.Extensions.Redis
         /// 获取Unix时间戳
         /// </summary>
         /// <returns></returns>
-        public static int GetUnixTime(this DateTime dt)
+        public static long GetUnixTime(this DateTime dt)
         {
             var ts = dt.ToUniversalTime() - new DateTime(1970, 1, 1, 0, 0, 0, 0);
-            return Convert.ToInt32(ts.TotalSeconds * 1000);
+            return Convert.ToInt64(ts.TotalSeconds * 1000);
         }
 
         /// <summary>
